requirements.txt
pandas
numpy
streamlit
python-binance
ccxt
ta
millify
sqlalchemy
